project cleanup: move all into single service for ease of development
