+++ .github/workflows/checks.yml
@@ -6,6 +6,9 @@ jobs:
   test:
     name: Test
     runs-on: ubuntu-latest
+    env:
+      CGO_CFLAGS_ALLOW: "-O -D__BLST_PORTABLE__"
+      CGO_CFLAGS: "-O -D__BLST_PORTABLE__"
     steps:
       - name: Set up Go 1.x
         uses: actions/setup-go@v2
@@ -22,6 +25,9 @@ jobs:
   lint:
     name: Lint
     runs-on: ubuntu-latest
+    env:
+      CGO_CFLAGS_ALLOW: "-O -D__BLST_PORTABLE__"
+      CGO_CFLAGS: "-O -D__BLST_PORTABLE__"
     steps:
       - name: Set up Go 1.x
         uses: actions/setup-go@v2
