+++ cmd/relay/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"flag"
 
-	"github.com/flashbots/boost-relay/apis/registervalidator"
+	"github.com/flashbots/boost-relay/server"
 	"github.com/sirupsen/logrus"
 )
 
@@ -11,7 +11,7 @@ var (
 	version = "dev" // is set during build process
 
 	// defaults
-	defaultListenAddr     = "localhost:9064"
+	defaultListenAddr     = "localhost:9062"
 	defaultBeaconEndpoint = "http://localhost:5052"
 
 	// cli flags
@@ -23,16 +23,16 @@ var log = logrus.WithField("module", "cmd/relay")
 
 func main() {
 	flag.Parse()
-	log.Printf("boost-relay %s [registervalidator-api]", version)
+	log.Printf("boost-relay %s [proposer-api]", version)
 
-	validatorService := registervalidator.NewBeaconClientValidatorService(*beaconEndpoint)
+	validatorService := server.NewBeaconClientValidatorService(*beaconEndpoint)
 	// TODO: should be done at the start of every epoch
 	err := validatorService.FetchValidators()
 	if err != nil {
 		log.WithError(err).Fatal("failed to fetch validators from beacon node")
 	}
 
-	srv, err := registervalidator.NewRegisterValidatorService(*listenAddr, validatorService, log)
+	srv, err := server.NewRelayService(*listenAddr, validatorService, log)
 	if err != nil {
 		log.WithError(err).Fatal("failed to create service")
 	}
