+++ server/beacon_client_test.go
@@ -1,4 +1,4 @@
-package registervalidator
+package server
 
 import (
 	"net/http"
